Rewrite elf loading related code for supporting uClibc ld.so.
load_icode is almost completely rewritten, now allowing more
program in ELF to be loaded even if they are in the same page.

However, bias support for ELF loading is currently removed,
because there seems to be no support for relocating.

uClibc ld.so's default load address is at 0x0 - 0x6000, so
I have to change the memory layout of amd64. It will be
changed back when dynamic load address is supported.

A large number of debugging statement is now active.