fix a bug in arch/arm/init/entry.S
We ought to load the SCTLR before write to it,
but here we directly write some value to it, which may overwrite its origin value.
Especially for SCTLR.XP, in armv7, this bit is kept to 1 and override
any write op. But qemu allow this, which causes cpu using legancy page
table format for armv5 and before.