fix bug when config global timer.
we should avoid the register write type such as:
```
do {
    outw();
}while(inw() != xxx)
```
That is because some bit in specific register is reserved or not
writable, so if you use such method, you'll be traped in this loop.
However, it seems that most(maybe even all) of registers in qemu is
writable and can be assigned to any value, so this bug may be hidded
when you are just using qemu.