fix bugs in early init code for ARM.
I've found that all the memory access is not corrent after enable mmu in
early init code in ./ucore/src/kern-ucore/arch/arm/init/entry.S. Then I
disable instruction and data cache in this stage. And the kernel work
fine.

In fact, If we've just create a whole new page table(which means that we
will then set the PDEP(ttb for ARM)), we should flush all cache or
disable them first. This is a common practice during boot stage.

In ucore, we should not enable cache until all the page table entries is
created in mmu_init() in pmm.c.